Move pins and color structs and add `StringArray` method to pins struct
