+++ pkg/picow/response.go
@@ -1,10 +1,7 @@
 package picow
 
-import "strconv"
-
 type (
 	None    any
-	Pins    []uint
 	Temp    float64
 	Version string
 )
@@ -12,16 +9,6 @@ type (
 type DiskUsage struct {
 	Used int `json:"used"`
 	Free int `json:"free"`
-}
-
-type Color []uint
-
-func (c *Color) StringArray() []string {
-	cS := make([]string, 0)
-	for _, n := range *c {
-		cS = append(cS, strconv.Itoa(int(n)))
-	}
-	return cS
 }
 
 type Response[T None | Pins | Color | Temp | Version] struct {
